Mini Server with caching capability - to handle create, read, write, delete file requests
The code implements a mini server with caching capability.
The server keeps on looking for new requests to handle. 
It can handle 4 type of requests for a file:
1. Create a file
2. Read from a file
3. Write to a file
4. Delete a file

It also implements a mini cache, which caches file name, FD related
information and lookups are made to the cache as needed.